requirements.txt
absl-py==1.4.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
flatbuffers==23.1.21
fonttools==4.38.0
ipython==8.10.0
jedi==0.18.2
kiwisolver==1.4.4
matplotlib==3.7.0
matplotlib-inline==0.1.6
mediapipe===dev
mido==1.2.10
numpy==1.24.2
opencv-contrib-python==4.7.0.72
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.37
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-rtmidi==1.4.9
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
vg==2.0.0
wcwidth==0.2.6
